# changes plugins reader and writer method