introduced interfaces.
```
// Message represents data accross plugins
type Message struct {
	Meta []byte // metadata
	Data []byte // actual data
}

// PluginReader is an interface for input plugins
type PluginReader interface {
	PluginRead() (msg *Message, err error)
}

// PluginWriter is an interface for output plugins
type PluginWriter interface {
	PluginWrite(msg *Message) (n int, err error)
}
```
### Benchmarks:
Google Compute Engine F1 micro
linux amd64

**THIS BRANCH:**
```
BenchmarkEmitter          401769              3180 ns/op             400 B/op         10 allocs/op
BenchmarkTCPOutput        169125              8729 ns/op             368 B/op          9 allocs/op
BenchmarkHTTPOutput        13290             94333 ns/op           10890 B/op         90 allocs/op
BenchmarkHTTPOutputTLS     10000            109845 ns/op           11008 B/op         93 allocs/op
```

**MASTER**
```
BenchmarkEmitter          386264              2865 ns/op             493 B/op         12 allocs/op
BenchmarkTCPOutput        187342              7763 ns/op             572 B/op         13 allocs/op
BenchmarkHTTPOutput        16778             77076 ns/op           11788 B/op         93 allocs/op
BenchmarkHTTPOutputTLS     12594            101574 ns/op           12094 B/op         96 allocs/op
```

The benchmarks run using a small payload, with a large payload the speed may seem to have improved!